+++ lib/spotify_io.dart
@@ -9,22 +9,11 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:http/http.dart' as http;
-import 'package:json_annotation/json_annotation.dart';
 
+import 'src/models/_models.dart';
 import 'src/models/token_request.dart';
 
-part 'spotify_io.g.dart';
-
-part 'src/models/paging.dart';
-part 'src/models/error.dart';
-part 'src/models/artist.dart';
-part 'src/models/followers.dart';
-part 'src/models/image.dart';
-part 'src/models/album.dart';
-part 'src/models/track.dart';
-part 'src/models/playlist.dart';
-part 'src/models/user.dart';
-part 'src/models/audio_feature.dart';
+export 'src/models/_models.dart';
 
 part 'src/endpoints/endpoint_base.dart';
 part 'src/endpoints/endpoint_paging.dart';
@@ -34,6 +23,7 @@ part 'src/endpoints/tracks.dart';
 part 'src/endpoints/playlists.dart';
 part 'src/endpoints/users.dart';
 part 'src/endpoints/audio_features.dart';
+part 'src/endpoints/search.dart';
 
 part 'src/spotify_credentials.dart';