+++ lib/src/models/paging.dart
@@ -1,15 +1,15 @@
 // Copyright (c) 2017, chances. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-part of spotify;
+part of spotify.models;
 
 typedef T ParserFunction<T>(dynamic object);
 
 Iterable<dynamic> itemsNativeFromJson(List<dynamic> json) => json;
 List<Map> itemsNativeToJson(Iterable<dynamic> items) => new List.from(items);
 
-@JsonSerializable()
-class Paging<T> extends Object with _$PagingSerializerMixin {
+@JsonSerializable(createToJson: false)
+class Paging<T> extends Object {
   Paging() {}
   factory Paging.fromJson(Map<String, dynamic> json) => _$PagingFromJson(json);
 
@@ -20,7 +20,8 @@ class Paging<T> extends Object with _$PagingSerializerMixin {
   ///
   /// Note this is the raw JSON value. Use a [Page]'s [Page.items] to get the
   /// requested data as a deserialized list.
-  @JsonKey(name: 'items', fromJson: itemsNativeFromJson, toJson: itemsNativeToJson)
+  @JsonKey(
+      name: 'items', fromJson: itemsNativeFromJson, toJson: itemsNativeToJson)
   Iterable<dynamic> itemsNative;
 
   /// The maximum number of items in the response (as set in the query or by