+++ lib/src/models/token_request.g.dart
@@ -7,32 +7,16 @@ part of spotify.token;
 // **************************************************************************
 
 TokenRequest _$TokenRequestFromJson(Map<String, dynamic> json) {
-  return new TokenRequest()..grantTypeString = json['grant_type'] as String;
+  return TokenRequest()..grantTypeString = json['grant_type'] as String;
 }
 
-abstract class _$TokenRequestSerializerMixin {
-  String get grantTypeString;
-  Map<String, dynamic> toJson() =>
-      <String, dynamic>{'grant_type': grantTypeString};
-}
+Map<String, dynamic> _$TokenRequestToJson(TokenRequest instance) =>
+    <String, dynamic>{'grant_type': instance.grantTypeString};
 
 ApiToken _$ApiTokenFromJson(Map<String, dynamic> json) {
-  return new ApiToken()
+  return ApiToken()
     ..accessToken = json['access_token'] as String
     ..refreshToken = json['refresh_token'] as String
     ..tokenType = json['token_type'] as String
     ..expiresIn = json['expires_in'] as int;
-}
-
-abstract class _$ApiTokenSerializerMixin {
-  String get accessToken;
-  String get refreshToken;
-  String get tokenType;
-  int get expiresIn;
-  Map<String, dynamic> toJson() => <String, dynamic>{
-        'access_token': accessToken,
-        'refresh_token': refreshToken,
-        'token_type': tokenType,
-        'expires_in': expiresIn
-      };
 }