+++ lib/src/spotify_mock.dart
@@ -8,7 +8,7 @@ class SpotifyApiMock extends SpotifyApiBase {
     var client = new MockClient();
     var response = await client.get(url, headers: headers);
 
-    return UTF8.decode(response.bodyBytes);
+    return utf8.decode(response.bodyBytes);
   }
 
   @override
@@ -16,14 +16,14 @@ class SpotifyApiMock extends SpotifyApiBase {
       String url, Map<String, String> headers, dynamic body) async {
     var client = new MockClient();
     var response = await client.post(url, headers: headers, body: body);
-    return UTF8.decode(response.bodyBytes);
+    return utf8.decode(response.bodyBytes);
   }
 
   @override
   Future<String> _putImpl(String url, Map<String, String> headers, body) async {
     var client = new MockClient();
     var response = await client.put(url, headers: headers, body: body);
-    return UTF8.decode(response.bodyBytes);
+    return utf8.decode(response.bodyBytes);
   }
 }
 
@@ -42,54 +42,59 @@ class MockClient implements http.BaseClient {
 
   @override
   void close() {
-    // TODO: implement close
+    throw "Not implemented";
   }
 
   @override
   Future<http.Response> delete(url, {Map<String, String> headers}) {
-    // TODO: implement delete
+    throw "Not implemented";
   }
 
   @override
-  Future<http.Response> get(url, {Map<String, String> headers}) async {
-    return new http.Response(_readPath(url), 200);
-  }
+  Future<http.Response> get(url, {Map<String, String> headers}) async =>
+      createSuccessResponse(_readPath(url));
 
   @override
   Future<http.Response> head(url, {Map<String, String> headers}) {
-    // TODO: implement head
+    throw "Not implemented";
   }
 
   @override
   Future<http.Response> patch(url,
       {Map<String, String> headers, body, Encoding encoding}) {
-    // TODO: implement patch
+    throw "Not implemented";
   }
 
   @override
   Future<http.Response> post(url,
-      {Map<String, String> headers, body, Encoding encoding}) async {
-    return new http.Response(_readPath(url), 200);
-  }
+          {Map<String, String> headers, body, Encoding encoding}) async =>
+      createSuccessResponse(_readPath(url));
 
   @override
   Future<http.Response> put(url,
       {Map<String, String> headers, body, Encoding encoding}) {
-    // TODO: implement put
+    throw "Not implemented";
   }
 
   @override
   Future<String> read(url, {Map<String, String> headers}) {
-    // TODO: implement read
+    throw "Not implemented";
   }
 
   @override
   Future<Uint8List> readBytes(url, {Map<String, String> headers}) {
-    // TODO: implement readBytes
+    throw "Not implemented";
   }
 
   @override
   Future<http.StreamedResponse> send(http.BaseRequest request) {
-    // TODO: implement send
+    throw "Not implemented";
+  }
+
+  http.Response createSuccessResponse(String body) {
+    /// necessary due to using Latin-1 encoding per default.
+    /// https://stackoverflow.com/questions/52990816/dart-json-encodedata-can-not-accept-other-language
+    return new http.Response(body, 200,
+        headers: {'Content-Type': 'application/json; charset=utf-8'});
   }
 }